chore(techinsights-addon): doesn't load the component for unsuported repositories
