+++ plugins/techdocs-module-addons-contrib/src/ReportIssue/ReportIssue.test.tsx
@@ -116,7 +116,7 @@ describe('ReportIssue', () => {
 
   it('renders gitlab link without exploding', async () => {
     byUrl.mockReturnValue({ type: 'gitlab' });
-    const { shadowRoot, getByText } =
+    const { shadowRoot, getByText, queryByTestId } =
       await TechDocsAddonTester.buildAddonsInTechDocs([
         <ReportIssue debounceTime={0} />,
       ])
@@ -164,6 +164,8 @@ describe('ReportIssue', () => {
     fireSelectionChangeEvent(window);
 
     await waitFor(() => {
+      expect(queryByTestId('report-issue-addon')).toBeInTheDocument();
+
       const link = getByText('Open new Gitlab issue');
       expect(link).toHaveAttribute(
         'href',
@@ -180,7 +182,7 @@ describe('ReportIssue', () => {
       body: options.selection.toString().trim(),
     });
 
-    const { shadowRoot, getByText } =
+    const { shadowRoot, getByText, queryByTestId } =
       await TechDocsAddonTester.buildAddonsInTechDocs([
         <ReportIssue debounceTime={0} templateBuilder={templateBuilder} />,
       ])
@@ -228,11 +230,58 @@ describe('ReportIssue', () => {
     fireSelectionChangeEvent(window);
 
     await waitFor(() => {
+      expect(queryByTestId('report-issue-addon')).toBeInTheDocument();
+
       const link = getByText('Open new Gitlab issue');
       expect(link).toHaveAttribute(
         'href',
         'https://gitlab.com/backstage/backstage/issues/new?issue[title]=Custom&issue[description]=his',
       );
+    });
+  });
+
+  it('does not render report issue link for unsupported repository type', async () => {
+    byUrl.mockReturnValue({ type: 'gerrit', resource: 'gerrit.example.com' });
+
+    const { shadowRoot, getByText, queryByTestId } =
+      await TechDocsAddonTester.buildAddonsInTechDocs([
+        <ReportIssue debounceTime={0} />,
+      ])
+        .withDom(
+          <html lang="en">
+            <head />
+            <body>
+              <div data-md-component="content">
+                <div data-md-component="main">
+                  <div className="md-content">
+                    <article>
+                      <a
+                        title="Edit this page"
+                        href="https://gerrit.example.com/admin/repos/edit/repo/my/repo/branch/refs/heads/master/file/docs/README.md"
+                      >
+                        Edit page
+                      </a>
+                    </article>
+                  </div>
+                </div>
+              </div>
+            </body>
+          </html>,
+        )
+        .withApis([[scmIntegrationsApiRef, { byUrl }]])
+        .renderWithEffects();
+
+    (shadowRoot as ShadowRoot & Pick<Document, 'getSelection'>).getSelection =
+      () => selection;
+
+    await waitFor(() => {
+      expect(getByText('Edit page')).toBeInTheDocument();
+    });
+
+    fireSelectionChangeEvent(window);
+
+    await waitFor(() => {
+      expect(queryByTestId('report-issue-addon')).not.toBeInTheDocument();
     });
   });
 });
