+++ plugins/techdocs-module-addons-contrib/src/ReportIssue/ReportIssue.tsx
@@ -14,39 +14,11 @@
  * limitations under the License.
  */
 
-import React, { useState, useEffect } from 'react';
-
-import { makeStyles, Portal, Paper } from '@material-ui/core';
-
-import { useGitTemplate, useGitRepository } from './hooks';
+import React from 'react';
+import { useGitRepository } from './hooks';
 import { ReportIssueTemplateBuilder } from './types';
-import {
-  PAGE_MAIN_CONTENT_SELECTOR,
-  PAGE_FEEDBACK_LINK_SELECTOR,
-  ADDON_FEEDBACK_CONTAINER_ID,
-  ADDON_FEEDBACK_CONTAINER_SELECTOR,
-} from './constants';
-import { IssueLink } from './IssueLink';
-
-import {
-  useShadowRootElements,
-  useShadowRootSelection,
-} from '@backstage/plugin-techdocs-react';
-
-const useStyles = makeStyles(theme => ({
-  root: {
-    transform: 'translate(-100%, -100%)',
-    position: 'absolute',
-    padding: theme.spacing(1),
-    zIndex: theme.zIndex.tooltip,
-    background: theme.palette.common.white,
-  },
-}));
-
-type Style = {
-  top: string;
-  left: string;
-};
+import { ADDON_ISSUE_REPO_TYPES_SUPPORTED } from './constants';
+import { ReportIssueAddonContent } from './ReportIssueContent';
 
 /**
  * Props customizing the <ReportIssue /> Addon.
@@ -67,101 +39,24 @@ export type ReportIssueProps = {
   templateBuilder?: ReportIssueTemplateBuilder;
 };
 
-/**
- * Show report issue button when text is highlighted
- */
 export const ReportIssueAddon = ({
   debounceTime = 500,
-  templateBuilder: buildTemplate,
+  templateBuilder,
 }: ReportIssueProps) => {
-  const classes = useStyles();
-  const [style, setStyle] = useState<Style>();
-
   const repository = useGitRepository();
 
-  const defaultTemplate = useGitTemplate(debounceTime);
-
-  const selection = useShadowRootSelection(debounceTime);
-
-  const [mainContent, feedbackLink] = useShadowRootElements([
-    PAGE_MAIN_CONTENT_SELECTOR,
-    PAGE_FEEDBACK_LINK_SELECTOR,
-  ]);
-
-  let [feedbackContainer] = useShadowRootElements([
-    ADDON_FEEDBACK_CONTAINER_SELECTOR,
-  ]);
-
-  if (feedbackLink) {
-    feedbackLink.style.display = 'none';
-  }
-
-  // calculates the position of the selected text to be able to set the position of the addon
-  useEffect(() => {
-    if (
-      // todo(backstage/techdocs-core) handle non-repo rendering
-      !repository ||
-      !selection ||
-      !selection?.containsNode(mainContent!, true) ||
-      selection?.containsNode(feedbackContainer!, true)
-    ) {
-      return;
-    }
-
-    const mainContentPosition = mainContent!.getBoundingClientRect();
-    const selectionPosition = selection.getRangeAt(0).getBoundingClientRect();
-
-    // Calculating the distance between the selection's top and the main content's top
-    const distanceFromTop = selectionPosition.top - mainContentPosition.top;
-    const minDistanceFromTop = 50;
-
-    // Defining a base value for 'top'
-    let top = distanceFromTop < minDistanceFromTop ? 101 : distanceFromTop - 16;
-
-    // Checking if the main content is off-screen towards the top
-    if (mainContentPosition.top < 0) {
-      const absMainContentTop = Math.abs(mainContentPosition.top);
-
-      // Adjusting 'top' if the selection is close to the top edge and the main content is off-screen
-      if (distanceFromTop - absMainContentTop < minDistanceFromTop) {
-        top += 89;
-      }
-    }
-
-    setStyle({
-      top: `${top}px`,
-      left: `${selectionPosition.left + selectionPosition.width / 2}px`,
-    });
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [selection, mainContent, feedbackContainer]);
-
   if (
-    !selection ||
     !repository ||
-    !['github', 'gitlab'].includes(repository.type)
-  )
+    !ADDON_ISSUE_REPO_TYPES_SUPPORTED.includes(repository.type)
+  ) {
     return null;
-
-  if (!feedbackContainer) {
-    feedbackContainer = document.createElement('div');
-    feedbackContainer.setAttribute('id', ADDON_FEEDBACK_CONTAINER_ID);
-    mainContent!.prepend(feedbackContainer);
   }
 
   return (
-    <Portal container={feedbackContainer}>
-      <Paper
-        data-testid="report-issue-addon"
-        className={classes.root}
-        style={style}
-      >
-        <IssueLink
-          repository={repository}
-          template={
-            buildTemplate ? buildTemplate({ selection }) : defaultTemplate
-          }
-        />
-      </Paper>
-    </Portal>
+    <ReportIssueAddonContent
+      debounceTime={debounceTime}
+      templateBuilder={templateBuilder}
+      repository={repository}
+    />
   );
 };
